multi_modal_rag_qa/requirements.txt
# Core utilities
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# PDF processing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# Table extraction (Camelot works with PyPDF2 & Ghostscript)
# Note: Requires Ghostscript to be installed separately
camelot-py[cv]>=0.11.0
PyPDF2>=3.0.0

# OCR (Requires Tesseract to be installed separately)
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Embeddings + Models
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# CLIP for image embeddings
git+https://github.com/openai/CLIP.git

# Vector DB (FAISS CPU version)
faiss-cpu>=1.7.4

# LLM API (Groq)
groq>=0.9.0

# Streamlit UI
streamlit>=1.28.0

# CLI formatting
rich>=13.0.0

# Optional helpers
tqdm>=4.65.0

# Fix compatibility issues
scikit-learn>=1.3.0
tabulate>=0.9.0
